Split FhirPathReader and FhirPathBuilder into modules
