+++ models/common/fhir-path-builder/AbstractBuilder.ts
@@ -1,0 +1,7 @@
+export class AbstractPathBuilder {
+  protected path: string
+
+  constructor(path: string) {
+    this.path = path
+  }
+}
